external status code of response template takes precedence
Order the response templates query by external status code so if you
have an update with an external status code the response template
matching that is always used. Previously with the database left to
determine the order it could pick a response template with a matching
state rather than a matching external status.